feat(mcp): extract remote proxy helpers into remote-proxy-lib
Move streamable HTTP proxy wiring and timeout fetch helpers into a
focused lib module with a thin re-export wrapper and targeted tests.
